pyproject.toml
[project]
name = "agent-auth"
version = "0.1.0"
description = "Agent authorization examples"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aioconsole>=0.8.1",
    "composio-client>=1.6.0",
    "google-genai>=1.28.0",
    "mcp>=1.12.3",
    "pydantic-ai-slim[openai]>=0.4.11",
    "python-dotenv>=1.1.1",
]
